test(algolia): add tests for algolia, fixes #2611 (#6952)
* test(algolia): add tests for algolia, fixes #2611

* chore(algolia): fix typo

* test(algolia): do it beter as unit tests

* refactor(algolia): update test descriptions and add test for `chunkAlgolia`

* refactor(algolia): run algolia index in own JS file

* refactor: rename test variables

+++ algolia.js
@@ -17,10 +17,13 @@ require('dotenv').config();
 const algoliasearch = require('algoliasearch');
 const byteof = require('byteof');
 const fs = require('fs');
+const path = require('path');
+
+const {defaultLocale, supportedLocales} = require('./shared/locale');
 
 const maxChunkSizeInBytes = 10000000; // 10,000,000
+const maxChunkSizeInMB = maxChunkSizeInBytes / 1000000;
 const maxItemSizeInBytes = 7000; // 7,000
-const {defaultLocale, supportedLocales} = require('./shared/locale');
 
 /**
  * Trim byte size of Algoia Collection Item.
@@ -55,15 +58,21 @@ const chunkAlgolia = (arr) => {
   let tempSizeInBytes = 0;
   let temp = [];
   for (const arrItem of arr) {
-    const current = trimBytes(arrItem);
-    const currentSizeInBytes = byteof(current);
+    const currentSizeInBytes = byteof(arrItem);
 
-    if (tempSizeInBytes + currentSizeInBytes < maxChunkSizeInBytes) {
-      temp.push(current);
+    if (currentSizeInBytes >= maxChunkSizeInBytes) {
+      throw new Error(
+        `${path.join(
+          arrItem.locale || '',
+          arrItem.url || '',
+        )} is >= than ${maxChunkSizeInMB} MB`,
+      );
+    } else if (tempSizeInBytes + currentSizeInBytes < maxChunkSizeInBytes) {
+      temp.push(arrItem);
       tempSizeInBytes += currentSizeInBytes;
     } else {
       chunked.push(temp);
-      temp = [current];
+      temp = [arrItem];
       tempSizeInBytes = currentSizeInBytes;
     }
   }
@@ -73,11 +82,6 @@ const chunkAlgolia = (arr) => {
 
 async function index() {
   const indexedOn = new Date();
-
-  if (!process.env.ALGOLIA_APP_ID || !process.env.ALGOLIA_API_KEY) {
-    console.warn('Missing Algolia environment variables, skipping indexing.');
-    return;
-  }
 
   const raw = fs.readFileSync('dist/pages.json', 'utf-8');
   /** @type {AlgoliaItem[]} */
@@ -105,49 +109,58 @@ async function index() {
       item.locales = [item.locale];
     }
     item.indexedOn = indexedOn.getTime();
-    return item;
+    return trimBytes(item);
   });
 
-  const chunkedAlgoliaData = chunkAlgolia(algoliaData);
-  const postsCount = algoliaData.length;
+  if (!process.env.ALGOLIA_APP_ID || !process.env.ALGOLIA_API_KEY) {
+    throw new Error(
+      'Missing Algolia environment variables, skipping indexing.',
+    );
+    // Ok, not a test, we got the keys, DO IT!
+  } else {
+    const chunkedAlgoliaData = chunkAlgolia(algoliaData);
+    const postsCount = algoliaData.length;
 
-  // @ts-ignore
-  const client = algoliasearch(
-    process.env.ALGOLIA_APP_ID,
-    process.env.ALGOLIA_API_KEY,
-  );
-  const index = client.initIndex('prod_web_dev');
+    // @ts-ignore
+    const client = algoliasearch(
+      process.env.ALGOLIA_APP_ID,
+      process.env.ALGOLIA_API_KEY,
+    );
+    const index = client.initIndex('prod_web_dev');
 
-  console.log(
-    `Indexing ${postsCount} articles amongst ${chunkedAlgoliaData.length} chunk(s).`,
-  );
+    console.log(
+      `Indexing ${postsCount} articles amongst ${chunkedAlgoliaData.length} chunk(s).`,
+    );
 
-  // When indexing data we mark these two fields as fields that can be filtered by.
-  await index.setSettings({
-    searchableAttributes: ['title', 'content', 'description'],
-    attributesForFaceting: ['locales', 'tags'],
-    customRanking: ['desc(priority)'],
-  });
-
-  // Update algolia index with new data
-  for (let i = 0; i < chunkedAlgoliaData.length; i++) {
-    await index.saveObjects(chunkedAlgoliaData[i], {
-      autoGenerateObjectIDIfNotExist: true,
+    // When indexing data we mark these two fields as fields that can be filtered by.
+    await index.setSettings({
+      searchableAttributes: ['title', 'content', 'description'],
+      attributesForFaceting: ['locales', 'tags'],
+      customRanking: ['desc(priority)'],
     });
+
+    // Update algolia index with new data
+    for (let i = 0; i < chunkedAlgoliaData.length; i++) {
+      await index.saveObjects(chunkedAlgoliaData[i], {
+        autoGenerateObjectIDIfNotExist: true,
+      });
+    }
+
+    console.log('Updated algolia data.');
+
+    console.log('Deleting old data no longer in algolia.json.');
+    await index.deleteBy({
+      filters: `indexedOn < ${indexedOn.getTime()}`,
+    });
+    console.log('Deleted old data.');
+    console.log('Done!');
   }
-
-  console.log('Updated algolia data.');
-
-  console.log('Deleting old data no longer in algolia.json.');
-  await index.deleteBy({
-    filters: `indexedOn < ${indexedOn.getTime()}`,
-  });
-  console.log('Deleted old data.');
-
-  console.log('Done!');
 }
 
-index().catch((err) => {
-  console.error(err);
-  throw err;
-});
+module.exports = {
+  chunkAlgolia,
+  index,
+  maxChunkSizeInBytes,
+  maxItemSizeInBytes,
+  trimBytes,
+};
